+++ CHANGELOG.md
@@ -12,8 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved `RedisPersistence` performance by re-using a couple of Redis connections instead of creating a new one each time.
 - Improved performance by adding a `JobExecutionPool` that keeps track of the instance's running job instead of fetching all jobs from the job persistence again.
 - Improved performance job selection: Loop over fetched job candidates and try to reserve a job instead of just trying the first job. This avoids re-fetching all jobs again after the first try.
-- Improved job timeout update by only updating timeout field and avoid using a transaction
-- Improved performance of fetching all jobs with a specific status (function `allJobsWithStatus`)
+- Improved job reservation for Redis by using a Lua script to avoid slower `transactionWithPreconditions` function
+- Improved performance by avoiding second check for stolen jobs, since the Lua script used for the job reservation is performed atomically.
+- Improved job timeout update by only updating timeout field and avoid using a transaction.
+- Improved performance of fetching all jobs with a specific status (function `allJobsWithStatus`).
 - Minor dependency updates
 
 ### Added